requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.5
decouple==0.0.7
defusedxml==0.8.0rc2
Django==4.2.5
django-allauth==0.58.1
django-ckeditor==6.7.0
django-crispy-forms==2.1
django-grappelli==3.0.8
django-js-asset==2.1.0
django-modeltranslation==0.18.11
django-rosetta==0.9.9
djangorestframework==3.14.0
drf-yasg==1.21.7
idna==3.4
inflection==0.5.1
joblib==1.3.2
numpy==1.26.2
oauthlib==3.2.2
packaging==23.2
Pillow==10.0.1
polib==1.2.0
psycopg==3.1.12
psycopg-binary==3.1.12
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
scikit-learn==1.3.2
scipy==1.11.3
social-auth-app-django==5.4.0
social-auth-core==4.5.0
sqlparse==0.4.4
threadpoolctl==3.2.0
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.7
